#18 [REFACTOR] chat 디렉토리 정리
